requirements.txt
fastapi
uvicorn[standard]
sqlmodel
apscheduler
python-dotenv
anthropic
google-api-python-client
google-auth
google-auth-oauthlib
notion-client
python-dateutil
mcp
requests